Implemented xml parser in Meta...Controller.java
This will increase the speed in which metadata will be accessed by the metadata overlay widget.  It takes only a subset of the metadata record and returns the subset in json form (note only OAI can be parsed).  Important assumption here is that the OAI metadata file is stored in the local temporary directory.