Reorganized Temporal and Metadata widgets
These widgets are now added as imports in livesearch.jsp and registered in solr.js.  This was done to fit into the 'mold' of the other ajax-solr widgets.  Note that the Metadata widget still has the exisiting problem of having to use the google maps api post document load with the use of a GLOBAL callback function.