Changed elements in the json data cart structure so that a dummy value can be placed as the 0th element of the array of datacart items.
This is a necessary evil for now.  The java JSON libraries cannot convert single element arrays (it creates a single object instead).  The datacart jquery template requires an array structure.  Thus we are forced to create a dummy element for the first index and just ignore it in the display.

Hopefully this glitched will be fixed sooner rather than later