WIP: Fixing the Datacart
This 'fix' is to allow users to filter out files based on search criteria.  Previously, ALL files were displayed...now they have a choice to display all files or just the files that satisfy the search constraints.

There are few more little fixes that need to be implemented before this can be merged.