Revert "No longer explicitly pulling in esg-search in code base"
Eventhough able to build without issue... there apparently is some
dependency that shows up upon deployment that yields a 500 error
because of what looks like still a lingering dependency on the search
jar.  For the moment we will curb this extracation process in lieu of
getting a working install to test the other features. John put in
place.

This reverts commit 257550fc10a8bc230a6a235f674c61fdd5cdd2e8.